Summary of Changes DevTools Option in Command-Line Arguments:
Added the --devtools argument to the command-line options for enabling Chrome DevTools Protocol.
Chrome Options for DevTools:

If the --devtools flag is set, Chrome is launched with the --auto-open-devtools-for-tabs argument to enable the developer tools.
JavaScriptChangeDetector Initialization:

Updated the initialization of JavaScriptChangeDetector to pass enable_devtools=args.devtools as a parameter.
Correct Invocation of fuzz_field():

Changed the invocation of fuzz_field() to fuzzer.fuzz_field(), which ensures that the method is called properly.